moved biz logic
